requirements.txt
numpy
paho-mqtt
sounddevice
scipy
tensorflow
